# server.go
package main

import (
	"fmt"
	"log"
	"net/http"
	"os"
	"strings"
	"sync"
)

const botToken = "8550220546:AAEVwWyWXhhUgJlMQaV8klu66qbzjzEFWNo"
const baseURL = "https://hr.monebakeryuz.uz"
const adminTgID int64 = 1066137436

// versionCheckURL — ilova (iOS/Android) yuklab olish linklarini beradigan
// tashqi endpoint. Login ma'lumotlari yuborilganda shu yerdan App Store /
// Play Store havolalari olinadi. appName'ni HR ilovasiga moslang.
const versionCheckURL = "https://version.uzaidev.uz/health?appName=hr_mobile_app"

// --- Telegram types ---

type TgUpdate struct {
	UpdateID      int64       `json:"update_id"`
	Message       *TgMessage  `json:"message"`
	CallbackQuery *TgCallback `json:"callback_query"`
}

type TgMessage struct {
	MessageID int64      `json:"message_id"`
	Chat      TgChat     `json:"chat"`
	From      *TgUser    `json:"from"`
	Text      string     `json:"text"`
	Contact   *TgContact `json:"contact"`
}

// TgContact — "Telefonni ulashish" tugmasi orqali kelgan kontakt. Matn o'rniga
// shu keladi, shuning uchun alohida qayta ishlanadi (aks holda bot raqamni
// tushunmay qayta-qayta so'rayveradi).
type TgContact struct {
	PhoneNumber string `json:"phone_number"`
	UserID      int64  `json:"user_id"`
}

type TgCallback struct {
	ID      string     `json:"id"`
	From    TgUser     `json:"from"`
	Message *TgMessage `json:"message"`
	Data    string     `json:"data"`
}

type TgChat struct {
	ID int64 `json:"id"`
}

type TgUser struct {
	ID        int64  `json:"id"`
	Username  string `json:"username"`
	FirstName string `json:"first_name"`
}

// --- Data types ---

type LangInfo struct {
	Til    string `json:"til"`
	Daraja string `json:"daraja"`
}

type Anketa struct {
	Lavozim string `json:"lavozim"`
	// Lavozimlar — anketada tanlangan barcha vakansiyalar. Bitta nomzod bir
	// nechta lavozimga ariza bera oladi; `Lavozim` esa birinchi tanlangani
	// (eski kod/mijozlar bilan moslik uchun) bo'lib qoladi.
	Lavozimlar      []string   `json:"lavozimlar"`
	Familiya        string     `json:"familiya"`
	Ism             string     `json:"ism"`
	Sharif          string     `json:"sharif"`
	TugilganSana    string     `json:"tugilgan_sana"`
	BoySm           int        `json:"boy_sm"`
	VaznKg          int        `json:"vazn_kg"`
	YashashManzili  string     `json:"yashash_manzili"`
	Moljal          string     `json:"moljal"`
	UmumiyTajriba   string     `json:"umumiy_tajriba"`
	ChetElTajribasi string     `json:"chet_el_tajribasi"`
	Malumot         string     `json:"malumot"`
	OilaviyHolat    string     `json:"oilaviy_holat"`
	Tillar          []LangInfo `json:"tillar"`
	Telefon         string     `json:"telefon"`
	Qoshimcha       string     `json:"qoshimcha"`
	Rasm            string     `json:"rasm"`
	// FaceRasm — Face ID apparati uchun yaqindan olingan yuz rasmi (base64).
	// Faqat ?start=oldWorker havolasidan kelgan anketalarda yuboriladi.
	FaceRasm        string     `json:"face_rasm"`
	TgUserID        int64      `json:"tg_user_id"`
	TgUsername      string     `json:"tg_username"`
	TgUsername2     string     `json:"tg_username2"`
	Source          string     `json:"source"`
}

type RezumeRow struct {
	ID      int64  `json:"id"`
	Lavozim string `json:"lavozim"`
	// Lavozimlar — rezumeda tanlangan barcha vakansiyalar (kamida bittasi —
	// `Lavozim`). Kategoriya bo'yicha filtr/hisob shu ro'yxat asosida ishlaydi.
	Lavozimlar      []string   `json:"lavozimlar"`
	Familiya        string     `json:"familiya"`
	Ism             string     `json:"ism"`
	Sharif          string     `json:"sharif"`
	TugilganSana    string     `json:"tugilgan_sana"`
	BoySm           int        `json:"boy_sm"`
	VaznKg          int        `json:"vazn_kg"`
	YashashManzili  string     `json:"yashash_manzili"`
	Moljal          string     `json:"moljal"`
	UmumiyTajriba   string     `json:"umumiy_tajriba"`
	ChetElTajribasi string     `json:"chet_el_tajribasi"`
	Malumot         string     `json:"malumot"`
	OilaviyHolat    string     `json:"oilaviy_holat"`
	Tillar          []LangInfo `json:"tillar"`
	Telefon         string     `json:"telefon"`
	Qoshimcha       string     `json:"qoshimcha"`
	RasmUrl         string     `json:"rasm_url"`
	FaceRasmUrl     string     `json:"face_rasm_url"`
	TgUserID        int64      `json:"tg_user_id"`
	TgUsername      string     `json:"tg_username"`
	TgUsername2     string     `json:"tg_username2"`
	Status          string         `json:"status"`
	StatusBy        int64          `json:"status_by"`
	StatusByName    string         `json:"status_by_name"`
	StatusVoiceUrl  string         `json:"status_voice_url"`
	Source          string         `json:"source"`
	CreatedAt       string         `json:"created_at"`
	Interviews      []InterviewRow `json:"interviews"`
}

type Category struct {
	ID        int64  `json:"id"`
	Name      string `json:"name"`
	IsActive  bool   `json:"is_active"`
	CreatedAt string `json:"created_at"`
}

type Branch struct {
	ID        int64   `json:"id"`
	Name      string  `json:"name"`
	Latitude  float64 `json:"latitude"`
	Longitude float64 `json:"longitude"`
	Address   string  `json:"address"`
	IsActive  bool    `json:"is_active"`
	CreatedAt string  `json:"created_at"`
}

type UserRow struct {
	ID           int64  `json:"id"`
	Username     string `json:"username"`
	Password     string `json:"password"`
	FullName     string `json:"full_name"`
	Role         string `json:"role"`
	// Roles — foydalanuvchining barcha rollari (masalan ["admin","ishchi_admin"]).
	// Eski `role` ustuni backward-compat uchun saqlanadi; bu ro'yxat asosiy.
	Roles        []string `json:"roles"`
	CanInterview bool     `json:"can_interview"`
	IsActive     bool   `json:"is_active"`
	BranchID     int64  `json:"branch_id"`
	RasmUrl      string `json:"rasm_url"`
	Telefon      string `json:"telefon"`
	RezumeID     int64  `json:"rezume_id"`
	CreatedAt    string `json:"created_at"`
}

type UserResponse struct {
	UserRow
	Categories       []Category `json:"categories"`
	IshchiCategories []Category `json:"ishchi_categories"`
	Branch           *Branch    `json:"branch"`
	// DeviceCount — bu foydalanuvchi login qilgan noyob qurilmalar soni
	// (user_devices jadvalidan). Faqat ko'rinish uchun. super_admin uchun 0
	// (uning qurilmasi yozilmaydi).
	DeviceCount int `json:"device_count"`
}

// UserDevice — foydalanuvchi login qilgan bitta qurilma. Dedup (user_id,
// device_id) bo'yicha: bir xil fizik qurilma ikki marta sanalmaydi.
type UserDevice struct {
	ID          int64  `json:"id"`
	DeviceID    string `json:"deviceId"`
	DeviceName  string `json:"deviceName"`
	Platform    string `json:"platform"`
	LastLoginAt string `json:"lastLoginAt"`
	CreatedAt   string `json:"createdAt"`
}

// --- Ishchi anketa types ---

type IshchiAnketa struct {
	Vakansiya    string `json:"vakansiya"`
	FIO          string `json:"fio"`
	TugilganSana string `json:"tugilgan_sana"`
	BoySm        int    `json:"boy_sm"`
	VaznKg       int    `json:"vazn_kg"`
	Manzil       string `json:"manzil"`
	Lang         string `json:"lang"`
	OilaviyHolat string `json:"oilaviy_holat"`
	Bolalar      string `json:"bolalar"`
	Tillar       string `json:"tillar"`
	Malumot      string `json:"malumot"`
	Grafik       string `json:"grafik"`
	Sudimlik     string `json:"sudimlik"`
	Haydovchilik string `json:"haydovchilik"`
	Telefon      string `json:"telefon"`
	Rasm         string `json:"rasm"`
	TgUserID     int64  `json:"tg_user_id"`
	TgUsername   string `json:"tg_username"`
	TgUsername2  string `json:"tg_username2"`
	Source       string `json:"source"`
}

type IshchiRow struct {
	ID             int64                `json:"id"`
	Vakansiya      string               `json:"vakansiya"`
	FIO            string               `json:"fio"`
	TugilganSana   string               `json:"tugilgan_sana"`
	BoySm          int                  `json:"boy_sm"`
	VaznKg         int                  `json:"vazn_kg"`
	Manzil         string               `json:"manzil"`
	Lang           string               `json:"lang"`
	OilaviyHolat   string               `json:"oilaviy_holat"`
	Bolalar        string               `json:"bolalar"`
	Tillar         string               `json:"tillar"`
	Malumot        string               `json:"malumot"`
	Grafik         string               `json:"grafik"`
	Sudimlik       string               `json:"sudimlik"`
	Haydovchilik   string               `json:"haydovchilik"`
	Telefon        string               `json:"telefon"`
	RasmUrl        string               `json:"rasm_url"`
	TgUserID       int64                `json:"tg_user_id"`
	TgUsername     string               `json:"tg_username"`
	TgUsername2    string               `json:"tg_username2"`
	Status         string               `json:"status"`
	StatusBy       int64                `json:"status_by"`
	StatusByName   string               `json:"status_by_name"`
	StatusVoiceUrl string               `json:"status_voice_url"`
	Source         string               `json:"source"`
	CreatedAt      string               `json:"created_at"`
	Interviews     []IshchiInterviewRow `json:"interviews"`
}

type IshchiInterviewRow struct {
	ID            int64   `json:"id"`
	IshchiID      int64   `json:"ishchi_id"`
	InvitedByID   int64   `json:"invited_by_id"`
	InvitedByName string  `json:"invited_by_name"`
	InterviewDate string  `json:"interview_date"`
	InterviewTime string  `json:"interview_time"`
	BranchID      int64   `json:"branch_id"`
	BranchName    string  `json:"branch_name"`
	BranchLat     float64 `json:"branch_lat"`
	BranchLng     float64 `json:"branch_lng"`
	Rating        int     `json:"rating"`
	RatingText    string  `json:"rating_text"`
	Comment       string  `json:"comment"`
	VoiceUrl      string  `json:"voice_url"`
	CreatedAt     string  `json:"created_at"`
	// Ishchi info
	IshchiFIO        string `json:"ishchi_fio"`
	IshchiVakansiya  string `json:"ishchi_vakansiya"`
	IshchiTelefon    string `json:"ishchi_telefon"`
	IshchiRasmUrl    string `json:"ishchi_rasm_url"`
	IshchiTgUsername string `json:"ishchi_tg_username"`
	IshchiTgUserID   int64  `json:"ishchi_tg_user_id"`
}

type InterviewRow struct {
	ID            int64   `json:"id"`
	RezumeID      int64   `json:"rezume_id"`
	InvitedByID   int64   `json:"invited_by_id"`
	InvitedByName string  `json:"invited_by_name"`
	InterviewDate string  `json:"interview_date"`
	InterviewTime string  `json:"interview_time"`
	BranchID      int64   `json:"branch_id"`
	BranchName    string  `json:"branch_name"`
	BranchLat     float64 `json:"branch_lat"`
	BranchLng     float64 `json:"branch_lng"`
	Rating        int     `json:"rating"`
	RatingText    string  `json:"rating_text"`
	Comment       string  `json:"comment"`
	VoiceUrl      string  `json:"voice_url"`
	CreatedAt     string  `json:"created_at"`
	// Rezume info
	RezumeFIO        string `json:"rezume_fio"`
	RezumeLavozim    string `json:"rezume_lavozim"`
	RezumeTelefon    string `json:"rezume_telefon"`
	RezumeRasmUrl    string `json:"rezume_rasm_url"`
	RezumeTgUsername string `json:"rezume_tg_username"`
	RezumeTgUserID   int64  `json:"rezume_tg_user_id"`
}

// --- State ---

var (
	userStates = make(map[int64]string)
	stateMu    sync.RWMutex

	// userSources — chatID -> /start <source> qiymati (sanitized).
	// Bot orqali kelgan foydalanuvchi qaysi reklama linkidan o'tganini eslab turadi.
	userSources       = make(map[int64]string)
	ishchiUserSources = make(map[int64]string)

	schedule   = make(map[string]int)
	scheduleMu sync.Mutex
)

var htmlPage string
var adminPage string
var demoPage string
var privacyPage string
var ishchiPage string

func main() {
	data, err := os.ReadFile("index.html")
	if err != nil {
		log.Fatal("index.html faylni o'qishda xato: ", err)
	}
	htmlPage = string(data)

	adminData, err := os.ReadFile("admin.html")
	if err != nil {
		log.Fatal("admin.html faylni o'qishda xato: ", err)
	}
	adminPage = string(adminData)

	demoData, err := os.ReadFile("demo.html")
	if err != nil {
		log.Fatal("demo.html faylni o'qishda xato: ", err)
	}
	demoPage = string(demoData)

	privacyData, err := os.ReadFile("privacy.html")
	if err != nil {
		log.Fatal("privacy.html faylni o'qishda xato: ", err)
	}
	privacyPage = string(privacyData)

	ishchiData, err := os.ReadFile("ishchi.html")
	if err != nil {
		log.Fatal("ishchi.html faylni o'qishda xato: ", err)
	}
	ishchiPage = string(ishchiData)

	os.MkdirAll("uploads", 0755)
	initDB()

	go startBotPolling()
	go startIshchiBotPolling()

	mux := http.NewServeMux()

	// Public — rezume yuborish
	mux.HandleFunc("/ws/rezumeler", handleWebSocket)
	mux.HandleFunc("POST /rezume", handleRezume)
	mux.HandleFunc("POST /ishchi-rezume", handleIshchiRezume)
	mux.HandleFunc("POST /api/report-error", handleReportError)
	mux.HandleFunc("GET /api/public/categories", handlePublicCategories)
	mux.HandleFunc("GET /api/public/ishchi-categories", handlePublicIshchiCategories)
	mux.HandleFunc("GET /api/public/rezume-by-phone", handleGetRezumeByPhone)
	mux.HandleFunc("GET /api/public/rezume-by-phone/{phone}", handleGetRezumeByPhone)

	// Auth
	mux.HandleFunc("POST /api/auth/login", handleLogin)
	mux.HandleFunc("POST /api/auth/logout", authRequired(handleLogout))
	mux.HandleFunc("GET /api/auth/me", authRequired(handleMe))

	// Rezume API (auth kerak)
	mux.HandleFunc("GET /api/rezumeler", authRequired(handleGetRezumeler))
	mux.HandleFunc("GET /api/rezumeler/counts", authRequired(handleGetRezumeCounts))
	mux.HandleFunc("GET /api/rezumeler/{id}", authRequired(handleGetRezume))
	mux.HandleFunc("PUT /api/rezumeler/{id}", authRequired(handleUpdateRezume))
	mux.HandleFunc("DELETE /api/rezumeler/{id}", authRequired(handleDeleteRezume))
	mux.HandleFunc("PATCH /api/rezumeler/{id}/status", authRequired(handleUpdateStatus))

	// Ishchi Anketa API (auth kerak)
	mux.HandleFunc("GET /api/ishchi-anketalar", authRequired(handleGetIshchiAnketalar))
	mux.HandleFunc("GET /api/ishchi-anketalar/{id}", authRequired(handleGetIshchiAnketa))
	mux.HandleFunc("PUT /api/ishchi-anketalar/{id}", authRequired(handleUpdateIshchiAnketa))
	mux.HandleFunc("DELETE /api/ishchi-anketalar/{id}", authRequired(handleDeleteIshchiAnketa))
	mux.HandleFunc("PATCH /api/ishchi-anketalar/{id}/status", authRequired(handleUpdateIshchiStatus))

	// Ishchi Interview API
	mux.HandleFunc("POST /api/ishchi-interviews", authRequired(handleCreateIshchiInterview))
	mux.HandleFunc("GET /api/ishchi-interviews", authRequired(handleGetIshchiInterviews))
	mux.HandleFunc("GET /api/ishchi-interviews/overdue", authRequired(handleGetIshchiOverdueInterviews))
	mux.HandleFunc("GET /api/ishchi-interviews/{id}", authRequired(handleGetIshchiInterview))
	mux.HandleFunc("PATCH /api/ishchi-interviews/{id}", authRequired(handleUpdateIshchiInterview))
	mux.HandleFunc("POST /api/ishchi-interviews/{id}/reschedule", authRequired(handleRescheduleIshchiInterview))
	mux.HandleFunc("DELETE /api/ishchi-interviews/{id}", authRequired(handleDeleteIshchiInterview))
	mux.HandleFunc("POST /api/ishchi-interviews/{id}/send-location", authRequired(handleSendIshchiInterviewLocation))

	// Interview API
	mux.HandleFunc("POST /api/interviews", authRequired(handleCreateInterview))
	mux.HandleFunc("GET /api/interviews", authRequired(handleGetInterviews))
	mux.HandleFunc("GET /api/interviews/overdue", authRequired(handleGetOverdueInterviews))
	mux.HandleFunc("GET /api/interviews/{id}", authRequired(handleGetInterview))
	mux.HandleFunc("PATCH /api/interviews/{id}", authRequired(handleUpdateInterview))
	mux.HandleFunc("POST /api/interviews/{id}/reschedule", authRequired(handleRescheduleInterview))
	mux.HandleFunc("DELETE /api/interviews/{id}", authRequired(handleDeleteInterview))
	mux.HandleFunc("POST /api/interviews/{id}/send-location", authRequired(handleSendInterviewLocation))

	// Stats API (super_admin) — reklama manbasi bo'yicha analitika
	mux.HandleFunc("GET /api/stats/sources", superAdminRequired(handleStatsRezumeSources))
	mux.HandleFunc("GET /api/stats/ishchi-sources", superAdminRequired(handleStatsIshchiSources))

	// User API (super_admin)
	mux.HandleFunc("POST /api/users", superAdminRequired(handleCreateUser))
	mux.HandleFunc("GET /api/users", superAdminRequired(handleGetUsers))
	mux.HandleFunc("PATCH /api/users/reorder", superAdminRequired(handleReorderUsers))
	mux.HandleFunc("GET /api/users/{id}", superAdminRequired(handleGetUserAPI))
	mux.HandleFunc("PATCH /api/users/{id}", superAdminRequired(handleUpdateUser))
	mux.HandleFunc("POST /api/users/{id}/link-rezume", superAdminRequired(handleLinkUserRezume))
	mux.HandleFunc("POST /api/users/{id}/replace-rezume", superAdminRequired(handleReplaceUserRezume))
	mux.HandleFunc("GET /api/users/{id}/devices", superAdminRequired(handleGetUserDevices))
	mux.HandleFunc("DELETE /api/users/{id}/devices/{deviceId}", superAdminRequired(handleDeleteUserDevice))
	mux.HandleFunc("DELETE /api/users/{id}", superAdminRequired(handleDeleteUser))
	// super_admin'dan boshqa barcha foydalanuvchilarga login ma'lumotlarini
	// bittada Telegram orqali yuborish — faqat super_admin.
	mux.HandleFunc("POST /api/users/send-all-credentials", superAdminRequired(handleSendAllCredentials))
	// Bitta foydalanuvchiga login ma'lumotlarini Telegram orqali yuborish.
	mux.HandleFunc("POST /api/users/{id}/send-credentials", superAdminRequired(handleSendCredentials))

	// Branch API
	mux.HandleFunc("POST /api/branches", superAdminRequired(handleCreateBranch))
	mux.HandleFunc("GET /api/branches", authRequired(handleGetBranches))
	mux.HandleFunc("PATCH /api/branches/{id}", superAdminRequired(handleUpdateBranch))
	mux.HandleFunc("DELETE /api/branches/{id}", superAdminRequired(handleDeleteBranch))

	// Category API
	mux.HandleFunc("POST /api/categories", superAdminRequired(handleCreateCategory))
	mux.HandleFunc("GET /api/categories", authRequired(handleGetCategories))
	mux.HandleFunc("PATCH /api/categories/{id}", superAdminRequired(handleUpdateCategory))
	mux.HandleFunc("DELETE /api/categories/{id}", superAdminRequired(handleDeleteCategory))

	// Ishchi Category API
	mux.HandleFunc("POST /api/ishchi-categories", superAdminRequired(handleCreateIshchiCategory))
	mux.HandleFunc("GET /api/ishchi-categories", authRequired(handleGetIshchiCategories))
	mux.HandleFunc("PATCH /api/ishchi-categories/{id}", superAdminRequired(handleUpdateIshchiCategory))
	mux.HandleFunc("DELETE /api/ishchi-categories/{id}", superAdminRequired(handleDeleteIshchiCategory))

	// Swagger
	mux.HandleFunc("GET /swagger", handleSwaggerUI)
	mux.HandleFunc("GET /swagger.json", handleSwaggerJSON)

	// Static
	mux.Handle("/uploads/", http.StripPrefix("/uploads/", http.FileServer(http.Dir("uploads"))))

	// Frontend
	mux.HandleFunc("GET /privacy", handlePrivacy)
	mux.HandleFunc("GET /demo", handleDemo)
	mux.HandleFunc("GET /admin", handleAdmin)
	mux.HandleFunc("/ishchi/", handleIshchi)
	mux.HandleFunc("/", handleRoot)

	handler := corsMiddleware(mux)

	log.Println("Server ishga tushdi: http://localhost:3333")
	log.Println("Swagger: http://localhost:3333/swagger")
	log.Fatal(http.ListenAndServe(":3333", handler))
}

func handleRoot(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	if strings.HasPrefix(r.URL.Path, "/ishchi/") || r.URL.Path == "/ishchi" {
		fmt.Fprint(w, ishchiPage)
		return
	}
	fmt.Fprint(w, htmlPage)
}

func handleAdmin(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	fmt.Fprint(w, adminPage)
}

func handlePrivacy(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	fmt.Fprint(w, privacyPage)
}

func handleDemo(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	fmt.Fprint(w, demoPage)
}

func handleIshchi(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	fmt.Fprint(w, ishchiPage)
}

func corsMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		w.Header().Set("Access-Control-Allow-Origin", "*")
		w.Header().Set("Access-Control-Allow-Methods", "GET, POST, PATCH, DELETE, OPTIONS")
		w.Header().Set("Access-Control-Allow-Headers", "Content-Type, Authorization")
		if r.Method == "OPTIONS" {
			w.WriteHeader(http.StatusOK)
			return
		}
		next.ServeHTTP(w, r)
	})
}

func handleSwaggerUI(w http.ResponseWriter, r *http.Request) {
	html := `<!DOCTYPE html>
<html lang="en">
<head>
    <meta charset="UTF-8">
    <title>HR Bot API - Swagger</title>
    <link rel="stylesheet" type="text/css" href="https://unpkg.com/swagger-ui-dist@5/swagger-ui.css">
    <style>html{box-sizing:border-box;overflow-y:scroll}*,*:before,*:after{box-sizing:inherit}body{margin:0;background:#fafafa}</style>
</head>
<body>
    <div id="swagger-ui"></div>
    <script src="https://unpkg.com/swagger-ui-dist@5/swagger-ui-bundle.js"></script>
    <script>
        window.onload = function() {
            SwaggerUIBundle({
                url: '/swagger.json',
                dom_id: '#swagger-ui',
                presets: [SwaggerUIBundle.presets.apis, SwaggerUIBundle.SwaggerUIStandalonePreset],
                layout: "BaseLayout"
            })
        }
    </script>
</body>
</html>`
	w.Header().Set("Content-Type", "text/html; charset=utf-8")
	fmt.Fprint(w, html)
}

func handleSwaggerJSON(w http.ResponseWriter, r *http.Request) {
	data, err := os.ReadFile("swagger.json")
	if err != nil {
		http.Error(w, "swagger.json topilmadi", http.StatusInternalServerError)
		return
	}
	w.Header().Set("Content-Type", "application/json")
	w.Header().Set("Access-Control-Allow-Origin", "*")
	w.Write(data)
}