# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HR Bot API",
    "description": "Mone Bakery HR rezume boshqaruv tizimi. Super admin userlar va kategoriyalarni boshqaradi. Har bir user o'ziga biriktirilgan kategoriyalarni ko'ra oladi.",
    "version": "2.0.0"
  },
  "servers": [
    {"url": "http://localhost:3333", "description": "Local development"},
    {"url": "https://hr.monebakeryuz.uz", "description": "Production"}
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "summary": "Tizimga kirish",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {"type": "string", "example": "admin"},
                  "password": {"type": "string", "example": "admin123"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Muvaffaqiyatli kirish",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {"type": "string", "example": "a1b2c3d4e5..."},
                    "user": {"$ref": "#/components/schemas/UserResponse"}
                  }
                }
              }
            }
          },
          "401": {"description": "Login yoki parol xato", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Tizimdan chiqish",
        "tags": ["Auth"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Chiqildi", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string", "example": "ok"}}}}}}
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Joriy foydalanuvchi ma'lumotlari",
        "tags": ["Auth"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Foydalanuvchi ma'lumotlari", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/UserResponse"}}}}
        }
      }
    },
    "/rezume": {
      "post": {
        "summary": "Rezume yuborish (public)",
        "description": "Yangi rezume yuboradi. Auth talab qilinmaydi. Rezume bazaga saqlanadi va Telegram guruhga yuboriladi.",
        "tags": ["Rezume"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/AnketaInput"},
              "example": {
                "lavozim": "Повар / Oshpaz",
                "familiya": "Yusupova",
                "ism": "Shoxista",
                "sharif": "Mamirovna",
                "tugilgan_sana": "28.11.1979",
                "boy_sm": 163,
                "vazn_kg": 50,
                "yashash_manzili": "Samarqand shaxar",
                "moljal": "Bozor",
                "umumiy_tajriba": "3 yil",
                "chet_el_tajribasi": "Yo'q",
                "malumot": "O'rta maxsus",
                "oilaviy_holat": "Oilaviy",
                "tillar": [
                  {"til": "Ingliz tili", "daraja": "bilmayman"},
                  {"til": "Rus tili", "daraja": "bilaman"},
                  {"til": "O'zbek tili", "daraja": "bilaman"}
                ],
                "telefon": "+998880640196",
                "qoshimcha": "",
                "rasm": "data:image/jpeg;base64,...",
                "tg_user_id": 123456789,
                "tg_username": "username"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Rezume yuborildi",
            "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}, "id": {"type": "integer"}}}}}
          }
        }
      }
    },
    "/api/rezumeler": {
      "get": {
        "summary": "Rezumeler ro'yxati",
        "description": "Foydalanuvchiga biriktirilgan kategoriyalar bo'yicha rezumeler. Super admin barchasini ko'radi.",
        "tags": ["Rezume"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "lavozim", "in": "query", "schema": {"type": "string"}, "description": "Lavozim bo'yicha filter (rezumening tanlangan lavozimlaridan biri mos kelsa chiqadi)"},
          {"name": "status", "in": "query", "schema": {"type": "string", "enum": ["yangi", "qabul", "rad"]}, "description": "Status bo'yicha filter"},
          {"name": "search", "in": "query", "schema": {"type": "string"}, "description": "Familiya, ism, telefon bo'yicha qidirish"},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20, "maximum": 100}}
        ],
        "responses": {
          "200": {
            "description": "Rezumeler ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"type": "array", "items": {"$ref": "#/components/schemas/RezumeRow"}},
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "limit": {"type": "integer"},
                    "pages": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rezumeler/{id}": {
      "get": {
        "summary": "Bitta rezume",
        "tags": ["Rezume"],
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Rezume", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/RezumeRow"}}}},
          "404": {"description": "Topilmadi", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      },
      "delete": {
        "summary": "Rezumeni o'chirish",
        "tags": ["Rezume"],
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "O'chirildi", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}}}}}}
        }
      }
    },
    "/api/rezumeler/{id}/status": {
      "patch": {
        "summary": "Rezume statusini yangilash (qabul/rad)",
        "tags": ["Rezume"],
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {"type": "string", "enum": ["yangi", "qabul", "rad"]}
                }
              },
              "example": {"status": "qabul"}
            }
          }
        },
        "responses": {
          "200": {"description": "Yangilandi", "content": {"application/json": {"schema": {"type": "object", "properties": {"status": {"type": "string"}}}}}}
        }
      }
    },
    "/api/interviews": {
      "post": {
        "summary": "Intervyuga chaqirish",
        "description": "Rezume egasini intervyuga chaqiradi. TG ga sana/vaqt bilan xabar yuboriladi. Kim chaqirgani saqlanadi.",
        "tags": ["Interviews"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rezume_id", "interview_date", "interview_time"],
                "properties": {
                  "rezume_id": {"type": "integer", "example": 1},
                  "interview_date": {"type": "string", "example": "28.03.2026", "description": "Intervyu sanasi"},
                  "interview_time": {"type": "string", "example": "09:00", "description": "Intervyu vaqti"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Interview yaratildi", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InterviewRow"}}}},
          "403": {"description": "Ruxsat yo'q", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Error"}}}}
        }
      },
      "get": {
        "summary": "Barcha intervyular",
        "tags": ["Interviews"],
        "security": [{"BearerAuth": []}],
        "parameters": [
          {"name": "rezume_id", "in": "query", "schema": {"type": "integer"}, "description": "Rezume bo'yicha filter"},
          {"name": "rating", "in": "query", "schema": {"type": "integer", "enum": [0, 1, 2, 3]}, "description": "0=kutilmoqda, 1=zo'r, 2=qoniqarli, 3=rad"},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 20}}
        ],
        "responses": {
          "200": {
            "description": "Intervyular ro'yxati",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {"type": "array", "items": {"$ref": "#/components/schemas/InterviewRow"}},
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "limit": {"type": "integer"},
                    "pages": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/interviews/{id}": {
      "get": {
        "summary": "Bitta intervyu",
        "tags": ["Interviews"],
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "responses": {
          "200": {"description": "Intervyu", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InterviewRow"}}}}
        }
      },
      "patch": {
        "summary": "Intervyu natijasini qo'yish",
        "description": "Intervyudan keyin baho (1=Zo'r, 2=Qoniqarli, 3=Qabul qilinmadi) va izoh qo'yish.",
        "tags": ["Interviews"],
        "security": [{"BearerAuth": []}],
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "integer"}}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rating"],
                "properties": {
                  "rating": {"type": "integer", "enum": [1, 2, 3], "description": "1=Zo'r, 2=Qoniqarli, 3=Qabul qilinmadi"},
                  "comment": {"type": "string", "example": "Yaxshi tajriba, mos keladi", "description": "Izoh/sabab"}
                }
              },
              "example": {"rating": 1, "comment": "Yaxshi tajriba, mos keladi"}
            }
          }
        },
        "responses": {
          "200": {"description": "Yangilandi", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/InterviewRow"}}}}
        }
      }
    },
...
    "/api/branches": {
      "post": {
        "summary": "Yangi filial yaratish",
        "tags": ["Branches"],
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object", "required": ["name"], "properties": {"name": {"type": "string", "example": "Gelion"}}}
            }
          }
        },
        "responses": {
          "200": {"description": "Yaratildi", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Branch"}}}}
        }
      },
      "get": {
        "summary": "Barcha filiallar",
        "tags": ["Branches"],
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {"description": "Filiallar ro'yxati", "content": {"application/json": {"schema": {"type": "array", "items": {"$ref": "#/components/schemas/Branch"}}}}}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Login qilib olingan tokenni kiriting"
      }
    },
    "schemas": {
...
      "Branch": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "name": {"type": "string", "example": "Gelion"},
          "is_active": {"type": "boolean"},
          "created_at": {"type": "string"}
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {"type": "integer"},
          "username": {"type": "string"},
          "full_name": {"type": "string"},
          "role": {"type": "string", "enum": ["admin", "super_admin"]},
          "can_interview": {"type": "boolean", "description": "Intervyuga chaqira oladi"},
          "is_active": {"type": "boolean"},
          "branch_id": {"type": "integer", "description": "Filial ID"},
          "created_at": {"type": "string"},
          "categories": {"type": "array", "items": {"$ref": "#/components/schemas/Category"}, "description": "Foydalanuvchiga biriktirilgan kategoriyalar"},
          "branch": {"$ref": "#/components/schemas/Branch", "nullable": true, "description": "Filial ma'lumotlari"}
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {"type": "string", "example": "Xato xabari"}
        }
      }
    }
  }
}